backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
numpy>=1.26.0
faster-whisper>=1.1.1
pyttsx3==2.90
soundfile>=0.12.1
